Moved the ReplicatedTestActorBase class and a minor change from PR feedback
